# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Juneflow API
  version: 0.1.0
  description: >-
    Construction ERP + Subscription SaaS (multi-tenant).
    REST, JSON, prefix /api/v1. Every request carries
    "Authorization: Bearer <jwt>"; the JWT carries company_id (tenant scope).
    Status changes go through action endpoints only, never direct PUT.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /project-types:
    get:
      tags: [master]
      summary: List project types (hierarchy[], modules{} — Full only)
      operationId: listProjectTypes
      parameters:
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/Page"
      responses:
        "200":
          $ref: "#/components/responses/EntityList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [master]
      summary: Create project type
      operationId: createProjectType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Entity"
      responses:
        "201":
          $ref: "#/components/responses/EntityCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /project-types/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [master]
      summary: Get project type by id
      operationId: getProjectType
      responses:
        "200":
          $ref: "#/components/responses/EntityOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [master]
      summary: Update project type
      operationId: updateProjectType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Entity"
      responses:
        "200":
          $ref: "#/components/responses/EntityOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT carries company_id = tenant scope (api-contract.md).
  parameters:
    Filter:
      name: filter
      in: query
      required: false
      schema:
        type: string
      description: Free-text/structured filter (GET /x?filter&page pattern).
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: 1-based page index (GET /x?filter&page pattern).
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing/invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found within the tenant scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    EntityOk:
      description: Resource (tenant-scoped by company_id from JWT).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Entity"
    EntityCreated:
      description: Created (AuditLog written by middleware).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Entity"
    EntityList:
      description: >-
        Paginated list envelope (B-014, tenant-scoped). data holds the page
        of rows; page/page_size/total are pagination metadata.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Paginated"
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Entity"
...
  schemas:
    Entity:
      type: object
      description: >-
        Opaque resource. Exact fields live in the data-dictionary schema tasks,
        not in this contract transcription — inventing fields here is forbidden
        (PLAN.md section 0). Modeled explicitly only where api-contract.md names
        fields (e.g. AuthLoginInput).
      additionalProperties: true
...
    Paginated:
      type: object
      description: >-
        Standard list envelope (B-014). Every list endpoint returns this
        wrapper: data is the page of rows (item type set per endpoint via
        allOf), the rest is pagination metadata. Tenant scope still applies.
      required: [data, page, page_size, total]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          minimum: 1
          example: 1
        page_size:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 137
    Error:
      type: object
      description: Standard error envelope.
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...